+++ services/client.js
@@ -1,7 +1,12 @@
+'use strict'
+
 const axios = require('axios')
 const rateLimit = require('axios-rate-limit')
 
 const {
+  JiraApiURL,
+  JiraUsername,
+  JiraAPIKey,
   togglAPIToken,
   togglBaseURL,
   tempoCloudURL,
@@ -20,6 +25,11 @@ const wrap = (axios) => rateLimit(axios, {
   maxRPS: 5
 })
 
+/**
+ * Client for Tempo Cloud
+ * @param {Object} options
+ * @return {AxiosInstance}
+ */
 module.exports.createTempoClient = (options = {}) => {
   return wrap(axios.create({
     baseURL: tempoCloudURL,
@@ -33,17 +43,38 @@ module.exports.createTempoClient = (options = {}) => {
 
 /**
  * Token buffering
+ * @param {string} username
  * @param {string} token
  * @return {string}
  */
-const bufferToken = (token) =>
-  Buffer.from(`${token}:api_token`).toString('base64')
+const bufferToken = (username, token) =>
+  Buffer.from(`${username}:${token}`).toString('base64')
 
+/**
+ * Client for Toggl
+ * @param {Object} options
+ * @return {AxiosInstance}
+ */
 module.exports.createTogglClient = (options = {}) => {
   return axios.create({
     baseURL: togglBaseURL,
     headers: {
-      Authorization: `Basic ${bufferToken(togglAPIToken)}`
+      Authorization: `Basic ${bufferToken(togglAPIToken, 'api_token')}`
+    },
+    ...options
+  })
+}
+
+/**
+ * Client for JIRA
+ * @param {Object} options
+ * @return {AxiosInstance}
+ */
+module.exports.createJiraClient = (options = {}) => {
+  return axios.create({
+    baseURL: JiraApiURL,
+    headers: {
+      Authorization: `Basic ${bufferToken(JiraUsername, JiraAPIKey)}`
     },
     ...options
   })